Move models into the `modules` package
