+++ src/main/java/nus/climods/model/module/ReadOnlyAcadYearModuleList.java
@@ -1,4 +1,4 @@
-package nus.climods.model;
+package nus.climods.model.module;
 
 import java.util.List;
